requirements.txt
click
Jinja2
flake8
python-dotenv
xmltodict
pytest
black
